진행정보 저장
사실상 디자인만 하면 끝
