+++ Interface/JSY8869/main.py
@@ -1,7 +1,7 @@
-import progress
+import main_page
 
 def main():
-    progress.PyShine_THREADS_APP()
+    main_page.PyShine_THREADS_APP()
 
 if __name__ == '__main__':
     main()
