--- Interface/JSY8869/test.txt
@@ -1,2 +1,0 @@
-dsfsadfreqwrwedsafsdaf
-dsafweqrewqrwqe
