--- Interface/JSY8869/wifi_ui.ui
@@ -1,91 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ui version="4.0">
- <class>Form</class>
- <widget class="QWidget" name="Form">
-  <property name="geometry">
-   <rect>
-    <x>0</x>
-    <y>0</y>
-    <width>571</width>
-    <height>479</height>
-   </rect>
-  </property>
-  <property name="windowTitle">
-   <string>Form</string>
-  </property>
-  <widget class="QTextEdit" name="wifi_name">
-   <property name="geometry">
-    <rect>
-     <x>100</x>
-     <y>110</y>
-     <width>321</width>
-     <height>31</height>
-    </rect>
-   </property>
-  </widget>
-  <widget class="QTextEdit" name="wifi_password">
-   <property name="geometry">
-    <rect>
-     <x>100</x>
-     <y>160</y>
-     <width>321</width>
-     <height>31</height>
-    </rect>
-   </property>
-  </widget>
-  <widget class="QLabel" name="label">
-   <property name="geometry">
-    <rect>
-     <x>30</x>
-     <y>120</y>
-     <width>64</width>
-     <height>15</height>
-    </rect>
-   </property>
-   <property name="text">
-    <string>wifi 이름</string>
-   </property>
-  </widget>
-  <widget class="QLabel" name="label_2">
-   <property name="geometry">
-    <rect>
-     <x>30</x>
-     <y>170</y>
-     <width>64</width>
-     <height>15</height>
-    </rect>
-   </property>
-   <property name="text">
-    <string>wifi 암호</string>
-   </property>
-  </widget>
-  <widget class="QPushButton" name="save_button">
-   <property name="geometry">
-    <rect>
-     <x>180</x>
-     <y>210</y>
-     <width>93</width>
-     <height>28</height>
-    </rect>
-   </property>
-   <property name="text">
-    <string>설정 저장</string>
-   </property>
-  </widget>
-  <widget class="QPushButton" name="home_button">
-   <property name="geometry">
-    <rect>
-     <x>290</x>
-     <y>210</y>
-     <width>93</width>
-     <height>28</height>
-    </rect>
-   </property>
-   <property name="text">
-    <string>돌아가기</string>
-   </property>
-  </widget>
- </widget>
- <resources/>
- <connections/>
-</ui>
